requirements.txt
streamlit==1.31.0
langchain==0.3.26
langchain-google-genai==2.0.9
langchain-community==0.3.27
langchain-core==0.3.80
langchain-text-splitters==0.3.8
pdfplumber==0.10.3
python-docx==1.1.0
faiss-cpu==1.9.0.post1
sentence-transformers==2.3.1
pypdfium2==5.1.0